requirements.txt
streamlit
google-generativeai
PyPDF2
langchain
python-dotenv
chromadb
faiss-cpu